Reorganise control files
