+++ src/components/Viewer.vue
@@ -125,9 +125,9 @@ import OpenSeadragon from 'openseadragon'
 import MetadataModal from '@/components/modals/Metadata'
 import HelpModal from '@/components/modals/Help'
 import BrowseModal from '@/components/modals/Browse'
-import ViewerControls from '@/components/controls/Viewer'
-import PanControls from '@/components/controls/Pan'
-import ZoomControls from '@/components/controls/Zoom'
+import ViewerControls from '@/components/types/items/Viewer'
+import PanControls from '@/components/types/items/Pan'
+import ZoomControls from '@/components/types/items/Zoom'
 import Sidebar from '@/components/sidebar/Sidebar'
 import SelectSidebarItem from '@/components/sidebar/items/Select'
 import TranscribeSidebarItem from '@/components/sidebar/items/Transcribe'
