+++ src/components/controls/Controls.vue
